Playback handler for client side (simple show beat text for demo) IN THIS COMMIT: - Fixed playback handler - Sends true to callbacks if whole beat
KNOWN BUGS:
- Gets a maximum recursion Error when activating sync fade and scene time. Probably a bug from the fader controller

KNOWN TODO'S:
- Finnish HTTP handler (FIXED A SIMPLE VERSION WITH THREADS, SLOW)
- Refactor BPM controller
- From JSON for scenes
- Strobe effects

- FIX Clients to display the color effect and receive the MSV signal

+++ LightStickController/Source/MsvController/msvcontroller.py
@@ -71,7 +71,7 @@ class SimpleMsvController(Widget, EventDispatcher):
 
     def sendSceneToServer(self, msg):
         pprint.pprint(msg)
-        msg["MSV_TIME"] = float(self.msvPosition)
+        msg["MSV_TIME"] = int(self.msvPosition)
         msg = json.dumps(msg)
         self.httpClient.postJson(jsonMessage=msg, url="/command")
 
@@ -83,7 +83,7 @@ class SimpleMsvController(Widget, EventDispatcher):
         print "Set new velocity"
         velocity = bpm / 60.0
         self.currentBPM = round(bpm, 1)
-        self.update(18, self.msvPosition, velocity, 0)
+        self.update(18, int(self.msvPosition), velocity, 0)
 
     def on_msvVelocity(self, *args):
         print args[1]
